requirements.txt
anthropic>=0.40.0
httpx>=0.27.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
sse-starlette>=2.1.3
deepgram-sdk>=3.7.0
rich>=13.8.0
